Move handler to index handlers
