+++ index-handlers/src/main/java/no/sikt/nva/nvi/index/utils/SearchConstants.java
@@ -3,8 +3,8 @@ package no.sikt.nva.nvi.index.utils;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import no.sikt.nva.nvi.index.aws.CandidateQuery;
-import no.sikt.nva.nvi.index.aws.CandidateQuery.QueryFilterType;
+import no.sikt.nva.nvi.index.query.CandidateQuery;
+import no.sikt.nva.nvi.index.query.CandidateQuery.QueryFilterType;
 import no.sikt.nva.nvi.index.model.search.CandidateSearchParameters;
 import nva.commons.core.Environment;
 import org.opensearch.client.opensearch._types.mapping.DoubleNumberProperty;
