+++ index-handlers/src/test/java/no/sikt/nva/nvi/index/aws/OpenSearchClientTest.java
@@ -1,20 +1,20 @@
-package no.sikt.nva.nvi.index;
+package no.sikt.nva.nvi.index.aws;
 
 import static no.sikt.nva.nvi.index.model.document.ApprovalStatus.NEW;
 import static no.sikt.nva.nvi.index.model.document.ApprovalStatus.PENDING;
 import static no.sikt.nva.nvi.index.model.document.ApprovalStatus.REJECTED;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.APPROVED_AGG;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.APPROVED_COLLABORATION_AGG;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.ASSIGNMENTS_AGG;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.COMPLETED_AGGREGATION_AGG;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.DISPUTED_AGG;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.NEW_AGG;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.NEW_COLLABORATION_AGG;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.PENDING_AGG;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.PENDING_COLLABORATION_AGG;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.REJECTED_AGG;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.REJECTED_COLLABORATION_AGG;
-import static no.sikt.nva.nvi.index.model.search.SearchAggregation.TOTAL_COUNT_AGGREGATION_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.APPROVED_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.APPROVED_COLLABORATION_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.ASSIGNMENTS_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.COMPLETED_AGGREGATION_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.DISPUTED_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.NEW_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.NEW_COLLABORATION_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.PENDING_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.PENDING_COLLABORATION_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.REJECTED_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.REJECTED_COLLABORATION_AGG;
+import static no.sikt.nva.nvi.index.query.SearchAggregation.TOTAL_COUNT_AGGREGATION_AGG;
 import static no.sikt.nva.nvi.test.TestUtils.randomBigDecimal;
 import static no.unit.nva.commons.json.JsonUtils.dtoObjectMapper;
 import static no.unit.nva.testutils.RandomDataGenerator.randomElement;
@@ -53,8 +53,7 @@ import java.util.UUID;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 import no.sikt.nva.nvi.common.service.model.GlobalApprovalStatus;
-import no.sikt.nva.nvi.index.aws.CandidateQuery.QueryFilterType;
-import no.sikt.nva.nvi.index.aws.OpenSearchClient;
+import no.sikt.nva.nvi.index.query.CandidateQuery.QueryFilterType;
 import no.sikt.nva.nvi.index.model.document.Approval;
 import no.sikt.nva.nvi.index.model.document.ApprovalStatus;
 import no.sikt.nva.nvi.index.model.document.Contributor;
@@ -66,7 +65,7 @@ import no.sikt.nva.nvi.index.model.document.PublicationDate;
 import no.sikt.nva.nvi.index.model.document.PublicationDetails;
 import no.sikt.nva.nvi.index.model.search.CandidateSearchParameters;
 import no.sikt.nva.nvi.index.model.search.OrderByFields;
-import no.sikt.nva.nvi.index.model.search.SearchAggregation;
+import no.sikt.nva.nvi.index.query.SearchAggregation;
 import no.sikt.nva.nvi.index.model.search.SearchResultParameters;
 import no.unit.nva.auth.CachedJwtProvider;
 import no.unit.nva.auth.CognitoAuthenticator;
