+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/utils/RequestUtil.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.nvi.utils;
+package no.sikt.nva.nvi.common.utils;
 
 import no.sikt.nva.nvi.common.service.model.Username;
 import nva.commons.apigateway.AccessRight;
@@ -6,7 +6,6 @@ import nva.commons.apigateway.RequestInfo;
 import nva.commons.apigateway.exceptions.UnauthorizedException;
 import nva.commons.core.JacocoGenerated;
 
-@JacocoGenerated
 public final class RequestUtil {
 
     private RequestUtil() {
